Fixed build and included all suites on the output

+++ CameraMouseSuiteCommon/User32.cs
@@ -48,7 +48,7 @@ namespace CameraMouseSuite
 
 		[DllImport("user32.dll")]
 		public static extern void mouse_event(int dwFlags, 
-			int dx, int dy, int dwData, long dwExtraInfo);
+			int dx, int dy, int dwData, int dwExtraInfo);
 
 		[DllImport("user32.dll")]
 		public static extern int GetSystemMetrics(int id);
